Split josh-command-map.ts into category-based modules #149
